# contracts/http/api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: CreatorFlow API Gateway
  version: 0.1.0
  description: Public API gateway untuk MVP CreatorFlow.
servers:
  - url: https://api.creatorflow.local
...
paths:
...
  /v1/trend/collect:
    post:
      tags: [Trend]
      summary: Manually collect a trend signal and create a lightweight digest
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualTrendCollectRequest"
      responses:
        "201":
          description: Trend signal collected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualTrendCollectResponse"
  /v1/trend/watchlist:
    get:
      tags: [Trend]
      summary: Get current trend watchlist for the active workspace
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Workspace trend watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendWatchlistResponse"
  /v1/trend/digests/latest:
    get:
      tags: [Trend]
      summary: Get the latest trend digest for the active workspace
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [daily, weekly]
      responses:
        "200":
          description: Latest trend digest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendDigestResponse"
  /v1/trend/recommendations/{recommendationId}/content-request:
    post:
      tags: [Trend]
      summary: Turn a trend recommendation into a content request
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: recommendationId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoteTrendToContentRequest"
      responses:
        "202":
          description: Content request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoteTrendToContentResponse"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ManualTrendCollectRequest:
      type: object
      properties:
        sourceCode:
          type: string
        sourceName:
          type: string
        signalType:
          type: string
          enum: [topic, product, angle, keyword]
        title:
          type: string
        recommendationType:
          type: string
          enum: [topic, product, angle, cta]
        recommendationText:
          type: string
        score:
          type: number
        period:
          type: string
          enum: [daily, weekly]
        summary:
          type: string
    ManualTrendCollectResponse:
      type: object
      required: [collectionRunId, snapshotId, recommendationId, digestId, status]
      properties:
        collectionRunId:
          type: string
          format: uuid
        snapshotId:
          type: string
          format: uuid
        recommendationId:
          type: string
          format: uuid
        digestId:
          type: string
          format: uuid
        status:
          type: string
          enum: [completed]
    TrendWatchlistItem:
      type: object
      required: [id, signalType, title, score, recommendationText, period]
      properties:
        id:
          type: string
          format: uuid
        signalType:
          type: string
          enum: [topic, product, angle, keyword]
        title:
          type: string
        score:
          type: number
        recommendationText:
          type: string
        period:
          type: string
          enum: [daily, weekly]
    TrendWatchlistResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TrendWatchlistItem"
    TrendDigestResponse:
      type: object
      required: [digestId, period, generatedAt, summary, items]
      properties:
        digestId:
          type: string
          format: uuid
        period:
          type: string
          enum: [daily, weekly]
        generatedAt:
          type: string
          format: date-time
        summary:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/TrendWatchlistItem"
    PromoteTrendToContentRequest:
      type: object
      properties:
        campaignId:
          type: string
          format: uuid
        titleOverride:
          type: string
        promptHint:
          type: string
    PromoteTrendToContentResponse:
      type: object
      required: [contentRequestId, status]
      properties:
        contentRequestId:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]